# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini Note RESTful API",
        "license": {
            "name": "Apache"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:2880/api/v1",
            "description": "API server development"
        },
        {
            "url": "https://mininote.orlangur.link/api/v1",
            "description": "API server production"
        }
    ],
    "paths": {
...
        "/users/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register user account",
                "description": "Registration user by email and password",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Password confirmation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login user account",
                "description": "Authorization user account by email and password",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/users/forgot": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Forgot user password",
                "description": "Recovery user password by email",
                "operationId": "forgot",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "restore_token": {
                                        "description": "Restore Token",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "New password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User profile",
                "description": "View user profile information",
                "operationId": "profile.read",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user account",
                "description": "Update user account",
                "operationId": "profile.update",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "description": "User avatar",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user account",
                "description": "Delete user account",
                "operationId": "profile.delete",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update password user account",
                "description": "Update password user account",
                "operationId": "password.update",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Password confirmation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/check": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Check exists user account",
                "description": "Check exists user account by email",
                "operationId": "check",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UniversalDTO": {
                "title": "UniversalDTO",
                "required": [],
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "type": "integer",
                                "example": "10"
                            },
                            "message": {
                                "type": "string",
                                "example": "Error message"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}